Refactor, 60 fps lock and timers changed to int for frames of animation

+++ Assets/Scripts/AttackScripts/AttackDataObject.cs
@@ -6,13 +6,10 @@ public class AttackDataObject : ScriptableObject
 {
     public int damage;
     public int startupFrames;//assuming 60 or 24 frames per second, this will be calculated later in the program frames/assuumedFramesOfAnimationPerSecond
-
     public int activeFrames;//assuming 60 or 24 frames per second, this will be calculated later in the program frames/assuumedFramesOfAnimationPerSecond
-
-    public float recoveryFrames;//assuming 60 or 24 frames per second, this will be calculated later in the program frames/assuumedFramesOfAnimationPerSecond
-
-    public float hitStun;
-    public float blockStun;
+    public int recoveryFrames;//assuming 60 or 24 frames per second, this will be calculated later in the program frames/assuumedFramesOfAnimationPerSecond
+    public int hitStun;
+    public int blockStun;
     public string animationName;
 
     [Header("Combo data")]
@@ -39,7 +36,4 @@ public class AttackDataObject : ScriptableObject
     public Vector3 origin; //possibly add more hurtboxes, for better customization
     public Vector3 size;
     public float radius;
-
-
-
 }
